feat: implement async support
